refactor: extract AuditLog columns and metadata dialog
Part of #103
