+++ src/pages/AuditLog.tsx
@@ -2,27 +2,9 @@ import React, { useState, useEffect, useCallback, useMemo } from 'react'
 import { useNavigate } from 'react-router-dom'
 import {
   Box,
-  Typography,
-  Avatar,
-  Chip,
-  IconButton,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogActions,
-  Button,
-  Tooltip,
-  Paper,
   Container,
-  useTheme,
 } from '@mui/material'
-import { Light as SyntaxHighlighter } from 'react-syntax-highlighter'
-import json from 'react-syntax-highlighter/dist/esm/languages/hljs/json'
-import { atomOneDark, atomOneLight } from 'react-syntax-highlighter/dist/esm/styles/hljs'
-
-SyntaxHighlighter.registerLanguage('json', json)
 import {
-  Visibility as VisibilityIcon,
   Person as PersonIcon,
   SwapHoriz as ProxyIcon,
   TrendingFlat as RedirectionIcon,
@@ -30,33 +12,25 @@ import {
   Block as DeadHostIcon,
   Security as AccessListIcon,
   VpnKey as CertificateIcon,
-  Description as AuditIcon,
   Add as AddIcon,
   Edit as EditIcon,
   Delete as DeleteIcon,
   PersonOutline as UserIcon,
-  Category as CategoryIcon,
-  PlayCircleOutline as ActionIcon,
-  Label as EntityIcon,
-  CalendarToday as DateIcon,
-  Settings as ActionsIcon,
 } from '@mui/icons-material'
-import { format } from 'date-fns'
 import { auditLogApi, AuditLogEntry } from '../api/auditLog'
 
 import PageHeader from '../components/PageHeader'
 import { DataTable } from '../components/DataTable'
-import { ResponsiveTableColumn, ColumnPriority } from '../components/DataTable/ResponsiveTypes'
 import { Filter } from '../components/DataTable/types'
 import { useToast } from '../contexts/ToastContext'
-import ActionChip from '../components/shared/ActionChip'
-import { NAVIGATION_CONFIG, NAVIGATION_COLORS } from '../constants/navigation'
+import { NAVIGATION_CONFIG } from '../constants/navigation'
 import { LAYOUT } from '../constants/layout'
 import { AUDIT_LOG_ROWS_PER_PAGE_OPTIONS } from '../constants/table'
+import { getAuditLogColumns, getObjectDisplayNameItems } from './AuditLog/auditLogColumns'
+import AuditLogMetaDialog from './AuditLog/AuditLogMetaDialog'
 
 const AuditLog = () => {
   const navigate = useNavigate()
-  const theme = useTheme()
   const [logs, setLogs] = useState<AuditLogEntry[]>([])
   const [loading, setLoading] = useState(true)
   const [error, setError] = useState<string | null>(null)
@@ -68,7 +42,7 @@ const AuditLog = () => {
     try {
       setLoading(true)
       setError(null)
-      const data = await auditLogApi.getAll({ 
+      const data = await auditLogApi.getAll({
         expand: ['user']
       })
       setLogs(data)
@@ -85,38 +59,15 @@ const AuditLog = () => {
     fetchAuditLogs()
   }, [fetchAuditLogs])
 
-  const handleViewMeta = (entry: AuditLogEntry) => {
+  const handleViewMeta = useCallback((entry: AuditLogEntry) => {
     setSelectedEntry(entry)
     setMetaDialogOpen(true)
-  }
+  }, [])
 
-  const handleCloseMetaDialog = () => {
+  const handleCloseMetaDialog = useCallback(() => {
     setMetaDialogOpen(false)
     setSelectedEntry(null)
-  }
-
-  const getObjectIcon = (objectType: string) => {
-    switch (objectType) {
-      case 'proxy-host':
-        return <ProxyIcon fontSize="small" sx={{ color: NAVIGATION_COLORS.success }} />
-      case 'redirection-host':
-        return <RedirectionIcon fontSize="small" sx={{ color: NAVIGATION_COLORS.warning }} />
-      case 'stream':
-      case 'stream-host':
-        return <StreamIcon fontSize="small" sx={{ color: NAVIGATION_COLORS.info }} />
-      case 'dead-host':
-        return <DeadHostIcon fontSize="small" sx={{ color: NAVIGATION_COLORS.danger }} />
-      case 'access-list':
-        return <AccessListIcon fontSize="small" sx={{ color: NAVIGATION_COLORS.primary }} />
-      case 'user':
-        return <PersonIcon fontSize="small" />
-      case 'certificate':
-        return <CertificateIcon fontSize="small" sx={{ color: NAVIGATION_COLORS.info }} />
-      default:
-        return null
-    }
-  }
-
+  }, [])
 
   const getObjectLink = (entry: AuditLogEntry): string => {
     switch (entry.object_type) {
@@ -148,226 +99,13 @@ const AuditLog = () => {
   }, [navigate])
 
   const getObjectDisplayName = useCallback((entry: AuditLogEntry) => {
-    const items: string[] = []
-    
-    switch (entry.object_type) {
-      case 'proxy-host':
-      case 'redirection-host':
-      case 'dead-host':
-        if (entry.meta.domain_names && entry.meta.domain_names.length > 0) {
-          items.push(...entry.meta.domain_names)
-        }
-        break
-      case 'stream':
-      case 'stream-host':
-        if (entry.meta.incoming_port) {
-          items.push(`Port ${entry.meta.incoming_port}`)
-        }
-        break
-      case 'access-list':
-      case 'user':
-        if (entry.meta.name) {
-          items.push(entry.meta.name)
-        }
-        break
-      case 'certificate':
-        if (entry.meta.provider === 'letsencrypt' && entry.meta.domain_names) {
-          items.push(...entry.meta.domain_names)
-        } else if (entry.meta.nice_name) {
-          items.push(entry.meta.nice_name)
-        }
-        break
-    }
-    
-    if (items.length === 0) {
-      return (
-        <Chip 
-          label={`#${entry.object_id || '?'}`}
-          size="small" 
-          sx={{ 
-            mx: 0.5, 
-            my: 0.25, 
-            cursor: 'pointer',
-            '& .MuiChip-label': {
-              px: 1.5,
-              py: 0.5
-            }
-          }}
-          onClick={() => handleChipClick(entry)}
-        />
-      )
-    }
-    
-    return items.map((item) => (
-      <Chip
-        key={item}
-        label={item} 
-        size="small" 
-        sx={{ 
-          mx: 0.5, 
-          my: 0.25, 
-          cursor: 'pointer',
-          '& .MuiChip-label': {
-            px: 1.5,
-            py: 0.5
-          }
-        }}
-        onClick={() => handleChipClick(entry)}
-      />
-    ))
+    return getObjectDisplayNameItems(entry, handleChipClick)
   }, [handleChipClick])
 
-  // Table column definitions with responsive priorities
-  const columns: ResponsiveTableColumn<AuditLogEntry>[] = useMemo(() => [
-    {
-      id: 'user',
-      label: 'User',
-      icon: <PersonIcon fontSize="small" />,
-      accessor: (entry) => entry.user.name,
-      sortable: true,
-      priority: 'P1' as ColumnPriority, // Essential - always visible
-      showInCard: true,
-      render: (_, entry) => (
-        <Box
-          sx={{
-            display: "flex",
-            alignItems: "center",
-            gap: 1.5
-          }}>
-          <Avatar
-            src={entry.user.avatar || '/images/default-avatar.jpg'}
-            sx={{ 
-              width: 40, 
-              height: 40,
-              border: entry.user.is_disabled ? '2px solid' : 'none',
-              borderColor: 'error.main'
-            }}
-          >
-            {entry.user.name.charAt(0).toUpperCase()}
-          </Avatar>
-          <Box>
-            <Typography
-              variant="body2"
-              sx={{
-                fontWeight: "medium",
-                textDecoration: entry.user.is_deleted ? 'line-through' : 'none'
-              }}>
-              {entry.user.name}
-            </Typography>
-            <Typography variant="caption" sx={{
-              color: "text.secondary"
-            }}>
-              {entry.user.email}
-            </Typography>
-          </Box>
-        </Box>
-      ),
-    },
-    {
-      id: 'object_type',
-      label: 'Type',
-      icon: <CategoryIcon fontSize="small" />,
-      accessor: (entry) => entry.object_type,
-      sortable: true,
-      priority: 'P2' as ColumnPriority, // Important - hidden on mobile
-      showInCard: true,
-      render: (_, entry) => (
-        <Box
-          sx={{
-            display: "flex",
-            alignItems: "center",
-            gap: 1
-          }}>
-          {getObjectIcon(entry.object_type)}
-          <Typography variant="body2" sx={{ textTransform: 'capitalize' }}>
-            {entry.object_type.replace('-', ' ')}
-          </Typography>
-        </Box>
-      ),
-    },
-    {
-      id: 'action',
-      label: 'Action',
-      icon: <ActionIcon fontSize="small" />,
-      accessor: (entry) => entry.action,
-      sortable: true,
-      priority: 'P1' as ColumnPriority, // Essential - always visible
-      showInCard: true,
-      render: (_, entry) => (
-        <ActionChip action={entry.action} />
-      ),
-    },
-    {
-      id: 'entity',
-      label: 'Entity',
-      icon: <EntityIcon fontSize="small" />,
-      accessor: (entry) => {
-        // For sorting, use the first domain name or name
-        if (entry.meta.domain_names?.[0]) return entry.meta.domain_names[0]
-        if (entry.meta.name) return entry.meta.name
-        if (entry.meta.nice_name) return entry.meta.nice_name
-        if (entry.meta.incoming_port) return `Port ${entry.meta.incoming_port}`
-        return `#${entry.object_id}`
-      },
-      priority: 'P1' as ColumnPriority, // Essential - always visible
-      showInCard: true,
-      mobileLabel: '',
-      render: (_, entry) => (
-        <Box
-          sx={{
-            display: "flex",
-            flexWrap: "wrap",
-            alignItems: "center"
-          }}>
-          {getObjectDisplayName(entry)}
-        </Box>
-      ),
-    },
-    {
-      id: 'created_on',
-      label: 'Date',
-      icon: <DateIcon fontSize="small" />,
-      accessor: (entry) => entry.created_on,
-      sortable: true,
-      priority: 'P2' as ColumnPriority, // Important - hidden on mobile
-      showInCard: true,
-      render: (date) => (
-        <Box>
-          <Typography variant="body2">
-            {format(new Date(date as string), 'MMM d, yyyy')}
-          </Typography>
-          <Typography variant="caption" sx={{
-            color: "text.secondary"
-          }}>
-            {format(new Date(date as string), 'h:mm a')}
-          </Typography>
-        </Box>
-      ),
-    },
-    {
-      id: 'actions',
-      label: 'Actions',
-      icon: <ActionsIcon fontSize="small" />,
-      align: 'right',
-      accessor: () => null,
-      priority: 'P1' as ColumnPriority, // Essential - always visible
-      showInCard: true,
-      render: (_, entry) => (
-        <Tooltip title="View metadata">
-          <IconButton
-            size="small"
-            aria-label="View audit entry metadata"
-            onClick={(e) => {
-              e.stopPropagation()
-              handleViewMeta(entry)
-            }}
-          >
-            <VisibilityIcon />
-          </IconButton>
-        </Tooltip>
-      ),
-    },
-  ], [getObjectDisplayName])
+  const columns = useMemo(
+    () => getAuditLogColumns(getObjectDisplayName, handleViewMeta),
+    [getObjectDisplayName, handleViewMeta]
+  )
 
   // Filter definitions
   const filters: Filter[] = useMemo(() => [
@@ -412,12 +150,8 @@ const AuditLog = () => {
   return (
     <Container maxWidth={false}>
       <title>Audit Log - NPMDeck</title>
-      <Box sx={{
-        py: 3
-      }}>
-        <Box sx={{
-          mb: 3
-        }}>
+      <Box sx={{ py: 3 }}>
+        <Box sx={{ mb: 3 }}>
           <PageHeader
             icon={React.createElement(NAVIGATION_CONFIG.auditLog.icon, { sx: { color: NAVIGATION_CONFIG.auditLog.color } })}
             title={NAVIGATION_CONFIG.auditLog.text}
@@ -444,114 +178,14 @@ const AuditLog = () => {
           compactBreakpoint={LAYOUT.COMPACT_BREAKPOINT}
         />
       </Box>
-      <Dialog
+
+      <AuditLogMetaDialog
         open={metaDialogOpen}
+        entry={selectedEntry}
         onClose={handleCloseMetaDialog}
-        maxWidth="md"
-        fullWidth
-      >
-        <DialogTitle>
-          <Box
-            sx={{
-              display: "flex",
-              alignItems: "center",
-              gap: 1
-            }}>
-            <AuditIcon sx={{ color: NAVIGATION_COLORS.dark }} />
-            Audit Log Details
-          </Box>
-        </DialogTitle>
-        <DialogContent dividers>
-          {selectedEntry && (
-            <Box>
-              <Box sx={{
-                mb: 3
-              }}>
-                <Typography variant="subtitle2" gutterBottom sx={{
-                  color: "text.secondary"
-                }}>
-                  Action Information
-                </Typography>
-                <Paper variant="outlined" sx={{ p: 2 }}>
-                  <Box
-                    sx={{
-                      display: "grid",
-                      gridTemplateColumns: "200px 1fr",
-                      gap: 1
-                    }}>
-                    <Typography variant="body2" sx={{
-                      color: "text.secondary"
-                    }}>User:</Typography>
-                    <Typography variant="body2">
-                      {selectedEntry.user.name} ({selectedEntry.user.email})
-                    </Typography>
-                    
-                    <Typography variant="body2" sx={{
-                      color: "text.secondary"
-                    }}>Action:</Typography>
-                    <ActionChip action={selectedEntry.action} />
-                    
-                    <Typography variant="body2" sx={{
-                      color: "text.secondary"
-                    }}>Object Type:</Typography>
-                    <Box
-                      sx={{
-                        display: "flex",
-                        alignItems: "center",
-                        gap: 1
-                      }}>
-                      {getObjectIcon(selectedEntry.object_type)}
-                      <Typography variant="body2" sx={{ textTransform: 'capitalize' }}>
-                        {selectedEntry.object_type.replace('-', ' ')}
-                      </Typography>
-                    </Box>
-                    
-                    <Typography variant="body2" sx={{
-                      color: "text.secondary"
-                    }}>Object ID:</Typography>
-                    <Typography variant="body2">{selectedEntry.object_id}</Typography>
-                    
-                    <Typography variant="body2" sx={{
-                      color: "text.secondary"
-                    }}>Date:</Typography>
-                    <Typography variant="body2">
-                      {format(new Date(selectedEntry.created_on), 'PPpp')}
-                    </Typography>
-                  </Box>
-                </Paper>
-              </Box>
-              
-              <Box>
-                <Typography variant="subtitle2" gutterBottom sx={{
-                  color: "text.secondary"
-                }}>
-                  Metadata
-                </Typography>
-                <Paper variant="outlined" sx={{ p: 0, backgroundColor: 'action.hover', overflow: 'hidden' }}>
-                  <SyntaxHighlighter
-                    language="json"
-                    style={theme.palette.mode === 'dark' ? atomOneDark : atomOneLight}
-                    customStyle={{
-                      margin: 0,
-                      padding: '16px',
-                      backgroundColor: 'transparent',
-                      fontSize: '0.875rem',
-                    }}
-                    wrapLongLines={true}
-                  >
-                    {JSON.stringify(selectedEntry.meta, null, 2)}
-                  </SyntaxHighlighter>
-                </Paper>
-              </Box>
-            </Box>
-          )}
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleCloseMetaDialog}>Close</Button>
-        </DialogActions>
-      </Dialog>
+      />
     </Container>
-  );
+  )
 }
 
 export default AuditLog
